requirements.txt
# Core dependencies
streamlit==1.29.0
google-generativeai==0.3.1
pandas==2.1.4
pillow==10.1.0
plotly==5.18.0
nltk==3.8.1
python-dotenv==1.0.0
requests==2.31.0

# SSL and security
cryptography==41.0.7
pyOpenSSL==23.2.0
urllib3==2.0.7

# Image generation dependencies
selenium==4.16.0
webdriver-manager==4.0.1 